Add support for variance in the `join` function
Summary:
This diff introduces support for type variance in the `join` function
of the type order.

One of the big decision points of this diff is what to return when
joining two instances of a parameter that the generic defines as
invariant, and when the instantiated types are unrelated; e.g
`join(List[int], List[str])`.

In this case, I decided to return `Type.Object` (AKA `Any`), as that
produces more readable error messages.

Despite that, the code changes are minimal - most of the code changes
relate to source code refactoring in the tests.

Reviewed By: sinancepel

Differential Revision: D13096041

fbshipit-source-id: 4fd6ce2fddd13a205e33ab8cbfbced3620a194ff