Add support for variance in the `meet` function
Summary:
This diff introduces support for type variance in the `meet` function
of the type order.

Similarly to what happened in previous diff in this stack, I took the
liberty to punt to Type.Bottom all instances of a parameter that the
generic defines as invariant, and when the instantiated types are
unrelated; e.g `meet(List[int], List[str])`.

Most of the code changes are localized in `meet`, with the usual
refactoring when it was clear that some functions were needed in
multiple places.

Reviewed By: sinancepel

Differential Revision: D13110837

fbshipit-source-id: 73ae295027372413771d314ee39591ece40a9e44