Make access path trees into ordinary domain
Summary:
Now that the AbstractDomain signature is powerful enough, we can get rid of the special casing of the AccessPathTree domain, and make it into a normal domain `AbstractTreeDomain`. The original domain is then a normal Map domain over roots to trees.

This diff just moves things around and renames.

Reviewed By: dkgi

Differential Revision: D13076565

fbshipit-source-id: fde8adfbbe8ce6bf4c5d2a65240ac1c3d90aebc3