Support a way to exclude files and directories from parsing (binary)
Summary:
Several users on Github issues have requested a way to permanently
exclude a set of files and directories from parsing. For example, see
issues #81, #92, #122. While the usage of the "do_not_check" flag
helps suppressing errors from a subset of unwanted files, it cannot
help when dealing with a set of problematic files; for example, files
with the wrong permission (#81), or files that make Pyre segfault (#119).

This diff adds support in the binary to exclude from parsing files and
directories that match a given regexp, or any in a list of regexps,
provided through a new optional, repeatable parameter in the binary:
`-exclude`.

Support will be added in the client in a subsequent diff.

Reviewed By: sinancepel

Differential Revision: D13210890

fbshipit-source-id: 95c80207decc5146144d6544100c4f4170f7e1df