Throw missing param errors in strict regardless of default value
Summary:
Pyre previously only warned on missing parameter annotations when a
default value was provided and no annotation was given; the annotation
suggestion was just the type of the default value. This is extremely limited,
and if we want strict mode to yell only about Anys introduced by the current
file, we need to be more broad about erroring here.

Reviewed By: dkgi

Differential Revision: D13195926

fbshipit-source-id: 0f87f08a880e2251da4bad8a7d76e0dd5bd34a53