Support single file strict by changing Any suppression
Summary:
Don't suppress parameter errors if they are due to the explicit
annotation being set to Any. Also start suppressing other due-to-any errors in
the body of the function, because any such errors with roots in this file will
throw on the signature where the Any is introduced, and other errors with roots
in other files (any function calls, etc.) should not error in strict mode.

(Missing return annotation, global annotation, and attribute annotation errors
are already appearing on untyped (even if typed to Any)
return/global/attributes in strict mode).

Reviewed By: sinancepel

Differential Revision: D13195924

fbshipit-source-id: 56ce14f136434f12805880f54adc5dec34749d24