Add API for retrieving typed AST for file by location
Summary:
Library functionality for taking in a builtin AST and augmenting it with types gathered from the pyre server.

This ends up working well regardless of whether we stick with location-based type lookup or not; it turns out that the builtin ast parsing happens under the hood of flake8, and plugins request this ast tree when initializing their AstChecker class. This tree is ideally treated as read-only, but modifications to the tree mean that the same modified tree is passed along to the next plugin that requests an AST to avoid re-parsing.

As a result, some plugins (ie. bugbear: https://github.com/PyCQA/flake8-bugbear/blob/master/bugbear.py#L29) have logic to re-parse the AST anyway in the case the tree received has been tampered with or deleted. Pyre can similarly offer interested plugins the ability to augment their given tree (location-based) or create a new tree entirely (if we migrate to ast-based), but either option will want to take in the old tree to at least check whether any work needs to be done at all.

For reference, output of a pyre query by file: P60293098

Reviewed By: sinancepel

Differential Revision: D12964040

fbshipit-source-id: c275f6c62f562ef4f89cb554a7842fe816380d06