Undefined Type errors are not due to analysis limitations
Summary: This doesn't change much in practice, but it makes no sense to me that we're checking whether an UndefinedType error is complaining on a specific type in the typeorder or not. If UndefinedType is thrown, it's never because of analysis limitations... it's because the type is not in the type order (if Type.Top is not in the typeorder, we're probably screwed in some other major way). We shouldn't suppress any of these.

Reviewed By: mrkmndz

Differential Revision: D13222657

fbshipit-source-id: d906399583fa1c16a69eab4a66f8cf401d1ff88f