Rename `do_not_check` to `ignore_all_errors`
Summary:
Several bugs were raised at times about the `do_not_check` feature,
stemming from the fact that it was unclear whether ignored files would
not be parsed at all, or the errors within would simply be thrown
away.

If diff renames `do_not_check` to `ignore_all_errors` in the hope
that the new name is clearer, and provides a deprecation workflow to
help users see the new change.

Reviewed By: shannonzhu

Differential Revision: D13221567

fbshipit-source-id: 4b71d606ad125f00bab902d206f7fbec5ec3c4a5