Pyre strict should not suggest unknown attribute annotations
Summary:
Kill the useless (and confusing) part of error messages. If we don't know what the type is, don't try to suggest one.

Note: Detail is also unnecessary here because the error itself is thrown at the location of the attribute definition, not where it's used (those are included in the evidence locations).

Reviewed By: dark

Differential Revision: D13226748

fbshipit-source-id: f1692ef67dbfa70200bce8a7b6d160029e4d833d