Expand namedtuples nested inside classes
Summary: In distillery there are instances of named tuples nested inside classes, then referred to as ParentClass.NamedTupleName in type signatures. We're yelling about undefined types when we come across these because we don't know about the Named Tuples. We are also throwing false positive missing attribute annotation errors on named tuple definitions nested in classes, which this will get rid of.

Reviewed By: sinancepel

Differential Revision: D13306383

fbshipit-source-id: d97c7ade9c186e5da484246a15e04653aad45a07