Check explicit annotations for undefined types
Summary: Need to refactor this check_annotation logic out for use to fix problem in D13267116 anyway. Realized we are not checking whether any explicit annotations living inside the function body are undefined or not. We're just letting them live on until they're checked at some point in a join or leq which will crash the analysis entirely later on. :/

Reviewed By: sinancepel

Differential Revision: D13268576

fbshipit-source-id: cfe96680792345d392ac2415bbb90dc98be1637d