Fixing generic constructors for list and dict
Summary:
This was quite a search for a one line change.  This problem was detected during the variance update, in a circumstance like this
```
def expect_list_of_tuples(x: List[typing.Tuple[str, str]]) -> None:
  pass
def foo(x: Mapping[str, int], y: str) -> None:
         codes = list(x.keys())
         number_list = [(y, code) for code in codes]
         # pyre-fixme[6]: Expected `List[typing.Tuple[str, str]]` but got `List[typing.Tuple[str, undefined]]`.
         expect_list_of_tuples(number_list)
```
This was being caused by the inferred type for `codes` being `List[undefined]`.

The root cause of this ended up being that in `Class.constructor`, the attribute was being instantiated with the type given by calling `Type.create` on the name of the class, which has a special case that replaces primitive `list` with `list[Any]`, which creates a callable without the type variable in the argument, but still has one in the return value (thanks to `Type.Callable.with_return_annotation`).  Thus this variable doesn't get populated in the call, leaving it as `List[undefined]`.

The same thing was also happening with the `dict` constructor.

Reviewed By: shannonzhu

Differential Revision: D13321883

fbshipit-source-id: b1054855ff7efdb55709f045a876d16e31efd736