Adding in class syntax for declaring a TypedDict
Summary:
This adds in mypy's alternate syntax for TypedDict declaration, only for one layer, not for arbitrary depth.
We have to do a little messing around with the qualification because this happens after the qualification step.  We could also move this before the qualification step and force a match on either TypedDict or mypy_extensions.TypedDict but would have no way of supporting `import ... as`.

I'm also making the other syntax's transformation more strict to make error messaging for mistakes better.

Reviewed By: shannonzhu

Differential Revision: D13245734

fbshipit-source-id: a1dfd7fbc5b4e3de8746c58a330ed9d2fdadd48d