Flush IDE state via LSP when a new file is opened
Summary:
A bug was reported where Nuclide would not remove errors from the UI
across updates to the Pyre backend. This was caused by Nuclide caching
results internally, if a "textDocument/publishDiagnostics" message is
not received.

This diff fixes the behaviour by making sure that the IDE flushes its
internal state of the opened files: this is done by sending out a
"textDocument/publishDiagnostics" message every time a file open
notification is received.

Reviewed By: sinancepel

Differential Revision: D13348838

fbshipit-source-id: f95d89e3eba16b240b1b53c24654247b2dfe26f9