Catch undefined type error in returns before analysis crashes
Summary:
At last, I think all of the most glaring issues with UndefinedType are fixed. :P

We're already throwing UndefinedType errors on bad return annotations, so we shouldn't crash analysis when the returns are checked. It also doesn't make much sense (results in tons of duplicate errors) if we throw the error at return time instead of at signature time.

Also adding in checking for casting to undefined type.

Reviewed By: dkgi

Differential Revision: D13269032

fbshipit-source-id: d6373e8aa5323b921d125a675d2e48e5dd30b424