Move advice for fixing incompatiblevar to detail
Summary:
Noticed doing the distillery upgrade that these error messages were disproportionately long, and unnecessarily so. This is the kind of line number information that goes in the detail for every other error type.

It's also not all that useful to display in the error because it usually points to the same line that the error is thrown on, anyway. :P

Reviewed By: sinancepel

Differential Revision: D13278143

fbshipit-source-id: e28caa38345432b7e440d85b73c34d8ac085d9cf