Fix behavioral subtyping check in case of nested types
Summary:
In the following code, Pyre would not recognize that `Derived` is
implementing `foo` correctly.

```
_T = TypeVar("_T")

class Base(Generic[_T]):
    def foo(self, x: Union[_T, str]) -> None:
        pass

class Derived(Base[int]):
    def foo(self, x: Union[int, str]) -> None:
        pass
```

This was due to the variables in `Base` not being properly replaced
with the constraints determined from the `Derived` class.

Fix the detection for parameters and return types.

Reviewed By: dkgi

Differential Revision: D13315356

fbshipit-source-id: f423f0c050a7f0414564ea00b7349a2496bea966