Store postcondition for entry node in cfg
Summary:
When creating a cfg out of a define node, the signature is not included as one of the statements; the resolution state after processing the signature is simply used as the precondition of the first body (block) cfg node.

Proposing to store this resolution state as the postcondition of the entry node (with an empty precondition, of course) in the shared resolution lookup. We know from testing the addition of postconditions that this has no perf impact, and I believe this is a lot cleaner than attempting to index into the block node and retrieve the precondition of the first statement there in order to know what the types are for the define signature statement.

See diff stacked on top of this for examples of why having this resolution is important and more useful than hacking together types for the signature from the AST itself.

Reviewed By: dkgi

Differential Revision: D13368720

fbshipit-source-id: 6aa372f11d7f6f9b1bec17f882c597a6df0dc233