Introduce postcondition to lookup logic for assign targets
Summary:
It makes sense in most cases that we use the precondition of statements to type expressions in that statement.

This does break down for assignments though, at least in terms of what behavior would be intuitive and useful.

```
x = 5
```
Currently returns an annotation for `int` at `5`, and `unknown` at `x`. Ideally we can use the postcondition for the target of assignments. That way, `x` also has type `int` here.

```
x = 5
x = int_to_str(x)
```

In this case, the first `x` will have type `int`, the `x` argument will have type `int`, and the `x` that is the target of the second assignment will have type `str`. IMO it would be extremely unintuitive and confusing for that second `x` target to still have type `int` on hover.

Unfortunately our visit module doesn't allow us to be flexible with what visit functions are called in what situations. I moved some things around in the visit function in a way that would allow overriding of pieces that can pull this off without duplicating code. Feedback on making this cleaner is very welcome.

Reviewed By: dark

Differential Revision: D13364545

fbshipit-source-id: b15c5e3429255db459f09502ac1535f9b332e190