Distinguish between explicit undefined types and uncaught crashes
Summary:
This seems like overly broad error suppression at first, but actually makes sense due to a huge prevalence of errors like:

a.py
```
def foo() -> int:
    ...
    x = b.ClassName.call()
    ...
    y = 5
    return y
```

b.py
```
   class ClassName:
        def call -> Foo:
             return Foo()

   class Foo:
         ... # There is an undefined type `Blah` used here
```

We will crash during our analysis of `a.py` and thrown an Undefined Type error on the line `def foo() -> int` saying `Undefined Type: Type Blah is undefined `.

As a user working in `a.py`, I have no idea what `Blah` is and in fact, it doesn't even exist in the entire `a.py` file. I can search around for it to no avail, and the error is pretty much meaningless, confusing, and seems like a glaring false positive.

In fact, the way to find the UndefinedType is to take a look at every single type-related assignment or operation inside the body of the function `foo`, and see if any of them called out to something that called out to something that may have supplied the undefined type that analysis is crashing on. When the Untracked Type exception is thrown in TypeOrder comparison operations, we have no idea where the type came from and when typecheck.ml catches that exception, can only assign the location of the UndefinedType error to be the signature line of the define that typecheck was analyzing oritinally.

I think these kinds of errors are very useful for debugging, but shouldn't surface outside of debug mode. We *should* throw visible user errors when we encounter an actual annotation that is Undefined; those errors are useful because they actually exist in the file and are something the user can change. We should be focusing on making sure all explicit annotations are fed through this check and errored against if they are undefined, and we should not be showing any far-fetched, typeorder-thrown undefined type errors in default or strict.

Reviewed By: sinancepel

Differential Revision: D13323120

fbshipit-source-id: 4167019bca8daf5b1a394b082265def4858a05e5