Use virtualenv pythonpath (#123)
Summary:
dark this is re: the comment you made in #117 about extending `--preserve-pythonpath`. When it's used, I include `sys.path` in the paths searched.  Please check the test output of  to verify the paths are as expected. With `PYTHONPATH=additional/:directories/`, it sets
```python
>>> configuration.search_path
[
  'additional/',
  'directories/',
  # vvv the contents of sys.path that are directories
  '',
   '/your/env/anaconda3/lib/python3.6',
   '/your/env/anaconda3/lib/python3.6/lib-dynload',
   '/your/env/anaconda3/lib/python3.6/site-packages',
   # ^^^
   'command/',
   'line/',
   'json/',
   'file/',
]
```

Assuming my understanding is correct and python searches `$PYTHONPATH` and then `sys.path` for packages, then this should fix #121 as well.
Pull Request resolved: https://github.com/facebook/pyre-check/pull/123

Differential Revision: D13413376

Pulled By: dark

fbshipit-source-id: 41de2eecb9175378e1e0d17649d60eaef5cabfcc