fix qualification for classes
Summary:
Scoping semantics in Python for classes are quite surprising - the outside scope of the class isn't captured for methods
in the class, even though things outside *that* scope, as well as parameter and return annotations are.

This means that we need to special-case the qualification of methods to accommodate this complexity, leading to a bit of code duplication, and some code getting moved around.

Open to better ways of solving this problem =)

Reviewed By: dkgi

Differential Revision: D13390959

fbshipit-source-id: e7d63702597b24bc102ba0dc7f9a2364b90b5247