Fix access lookup after global assignment to attributes
Summary:
We register globals that are assigned at toplevel, so when they are used in subsequent logic (in defines) we know what the local annotation is. However, at environment building time, we don't know what the original annotation was, and so we leave it at Type.Top in the case that the assignment we are processing wasn't explicitly annotated.

The following example:

      class Foo:
        a: typing.Any
      Foo.a = 1

Used to throw: Missing attribute annotation [4]: Attribute `a` of class `Foo` has type `int` but no type is specified.

And

            class Foo:
              a: str = ""
            Foo.a = 1

threw: Missing attribute annotation [4]: Attribute `a` of class `Foo` has type `typing.Union[int, str]` but no type is specified.

This is because `Foo.a` was registered in globals as having annotation (local: int, original: top). The local int is actually necessary in order to do something like:

```
   Foo.a = 1
   def foo() -> int:
      return Foo.a
```
since it is a globally accessible value after all. However, when we look up the annotation of `Foo.a` in access fold, we should understand that the original annotation comes from the annotation we find in the class definition.

PS. Follow-up problems this dug up:
T37918591

Reviewed By: sinancepel

Differential Revision: D13410722

fbshipit-source-id: 20b7ed5c6896e9093b207a8f167a57d2ee1dc83f