Never do global lookup of instance attributes in access fold
Summary:
This is also a rfc.
Not sure if this can be done more intelligently by not adding these to the environment in the first place... but all we have is a chain of identifiers and no way to look up whether what we are looking at is an instance attribute or not.

For more context, the test added in this diff used to fail because it threw:
Missing attribute annotation [4]: Attribute a` of class `Foo` has type `int` but type `Any` is specified.
Incompatible return type [7]: Expected `bool` but got `int`.

Reviewed By: sinancepel

Differential Revision: D13418054

fbshipit-source-id: 14a6dcc277fe3f2f7de6f3f3a7628b4165892c24