Correctly determine dependent files when processing file updates
Summary:
There were cases where the server would not run a type check over some
files that needed re-checking; this would cause the server and
`pyre check` to go out of sync. The problem was concentrated on
intermediate dependencies, not propagating the update all the way to
second-level dependencies, and was discovered while working on starred
imports.

The fix is twofold.

1. When a given file is updated, don't just recheck dependent files,
but update them in the environment too. This is needed because
dependent files might define classes or methods that depend on the
updated definition. This will implicitly cause second-level
dependencies to be updated too in a subsequent iteration, since the
first-level dependencies will pull them in as their "deferred
requests".

2. Use heuristics to determine if a starred import should cause a
dependent to be re-checked.

Reviewed By: sinancepel

Differential Revision: D12929087

fbshipit-source-id: 6de643836b855ee347e0e04baa5eac9d47457a20