undeclared vincit omnia
Summary:
Sometimes we would lose track that a given annotation is undeclared in
a code branch, because a parallel Top would just overwrite it. This
would hide potentially undeclared variables, if one of the branches
happens to assign `unknown` to that variable.

Fix the issue by always performing the union between `top` and
`undeclared` if they need to be joined.

Reviewed By: shannonzhu

Differential Revision: D13428861

fbshipit-source-id: dcae8dc29868ad689062dd1767a9cf89b0b13f01