Set filter dir to location of TARGETS when no local config exists
Summary:
When a local configuration did not exist, we were setting filter_directory to be equivalent to whatever the final analysis directory (aka link tree aka source directory) is. For example, if the analysis directory ended up being buck-out/gen/.../#link-tree, then we'd simply set the filter directory to the same. However, when we actually do filter directories in the back end, we resolve all symlinks before comparing to the filter directory passed in. This would result in calls to pyre with targets checking 0 files.

While going in to change this, I noticed that filter directories should be determined (and depend only on) the calculation we are already doing to determine what the final analysis directory is. We should be figuring it out there and linking the two to use later, not ad-hoc determining it at the end. Did some refactoring to clean up the logic surrounding this, but more cleaning and refactoring can definitely be done but wanted to stack on top to keep this diff simpler.

~31/7.6 runs (excluding sandcastle) passed in targets explicitly in the past day
Scuba: https://fburl.com/scuba/rg2vvi0b

Some follow-ups:
 - Want to clean up references to local_configuration_root. This is being passed around as an extra attribute attached to arguments for some reason, diluting the source of truth and making the code less readable
- Perhaps some cleaning of other confusing terminology (ex. arguments.target is a list of targets, not one target)
- Still need to deal with buck-out directories being passed in as a source directory itself. This is necessary if we have just built the buck target and don't want to go through the work of doing it again. There are two options here, will elaborate further on another diff.

Reviewed By: dkgi

Differential Revision: D13426963

fbshipit-source-id: e8a8d04875fb234f9fd32c3488e6dbd6214547dd