Refactor away local_configuration_directory field of client arguments
Summary:
The concept of a `local_configuration_directory` floating around in arguments was confusing to me - especially because arguments already contain a `local_configuration` (passed in with the `-l` flag) that is a separate concept, and `local_configuration_directory` doesn't correspond to anything that can be passed in in arguments.

Turns out this field is being populated in `switch_root`, which searches upward from the cwd to find and use a local or global configuration that is in a parent directory of the current one. This functionality is important - but we can kill the new field and keep a single source of truth because
1. If a local configuration was explicitly passed in as an argument to pyre, we should use that one
2. If a local config wasn't passed in explicitly and we find a valid local config looking upwards in switch root, then we should just fill in the local config space
3. If targets or source directories were passed in as arguments to pyre, they already take precedence over the contents of any local configuration we know about.

Reviewed By: sinancepel

Differential Revision: D13436714

fbshipit-source-id: 698f09d0d9fe9d7a0e76a657893efabd450f2c07