Distinguish between real directories and constructed analysis directories
Summary:
I know this has been back and forth, but I dare suggest this change might encompass both the reason we wanted to move to analysis_directory *and* the reason we pulled back from exposing that to the user in the end.

 ---

Motivation:

- Found it surprisingly hard to hold state in my head when reading existing logic, since it contains references to both source and analysis directories, and all our documentation is named source only while the corresponding logic is mostly named analysis.
- More confusion comes when we are creating our own analysis directories out of the input analysis directories. We end up with logic that does stuff like:
```
analysis_directories = _resolve_analysis_directories(arguments.analysis_directories)
if len(analysis_directories) = 1:
    analysis_directory = AnalysisDirectory(analysis_directories[0])
else:
    analysis_directory = SharedAnalysisDirectory(analysis_directories)
```
We're dealing with several "kinds" of analysis directories here. The ones the user passes in and the ones we are deriving. In order to avoid this, these helper functions need to be named something else, but either way it's confusing because we are conflating two concepts here.

----

Definition of source/analysis directory that this diff assumes:

Source directory = some directory that maps to a real tangible external directory of source files. The user passes this in - to them, they never need to know what an analysis directory is because they just want to run pyre on their directories of source files

Analysis directory = what pyre ends up actually analyzing. This is often some glommed combination directory that we generate in order to be eden-compatible or to dedup dependencies or pull in multiple buck link trees into one. It's not really accurate to call the end result a source directory because it doesn't map to real source files.

 ---

My understanding is the initial move away from using `source directory` is because of the definition of `analysis directory` above, and the reason we stopped halfway is because this is confusing to users who really are passing in source directories.

Hopefully this makes sense as a distinction - but very happy to take suggestions and open to abandoning this if I misunderstood and this still makes no sense to everyone else (though I'd still want to distinguish by some other name the input directories and the generated objects we create in the backend).

Reviewed By: dkgi

Differential Revision: D13438031

fbshipit-source-id: 6d98fcd1e04ff302a312a16b9446bd3cd722974f