Rename arguments.target to reflect plurality of data
Summary:
This renaming is already done for source directories in the big rename diff below this.

However, targets too are passed in with the label `--target`, but multiple `--target` labels can be used in one call to pyre and the argparse `append` action creates a list out of all of them. Regardless of how many there are, `arguments.target` is always a *list* of targets, and this makes it awfully difficult to understand and is bug-prone when editing later on when we are messing with `arguments`.

Reviewed By: dark

Differential Revision: D13438041

fbshipit-source-id: 73d3ada7b5db540edef2dbe7b95465e0d1fcd03f