Start and restart should always rebuild buck targets
Summary:
Adding or removing files will cause buck link trees to become stale while still exisitng in a 1:1 mapping with the targets pyre is running on. In order to support incremental (/ generally to avoid doing duplicate work), shelling to pyre's buck integration will check if the appropriate link trees exist already and only rebuild the targets if they don't. We should assume link trees are stale when `pyre restart` or `pyre start` are called explicitly.

Also tried to make the purpose of the `--build` flag a little clearer.

Reviewed By: dark

Differential Revision: D13452739

fbshipit-source-id: 2ed723f5b3e2c62ed4d0a5d33706edce83d65f9d