Fix inference of callable methods
Summary:
The core problem this change addresses is that the callable attribute being inferred depends on the class it comes from, but we weren't respecting this fact,
instead relying on heuristics of having access to the calling class in the access chain. The change makes it so that if I call `c.f`, `f` already has the self argument
accounted for, which models the runtime more accurately than trying to introduce the class when applying the function to arguments.

Reviewed By: dkgi

Differential Revision: D13425923

fbshipit-source-id: 88e34c19fa15f6208b0d54a627d61f1d84671a0e