show types which have a Top somewhere in them in lookups
Summary:
We don't want to show things that are blatantly unknown, but I'd argue that it makes sense to show annotations for
obejcts where we know the basic shape (List, Dict, ...). A future change will make the printing nicer, but wanted to keep this as is
to avoid poking the type checking tests beast.

Reviewed By: dark

Differential Revision: D13456877

fbshipit-source-id: 29f0fc106f1be53494cf7372221269ff8e3eea8e