Save error code metadata
Summary:
We need to start saving extra meta data per analysis. To do so, we interpret the --save-results-to flag as a directory and store
  XX-output.json
  XX-metadata.json

for each analysis XX that ran.

The metadata.json can be used by the analysis to dump extra info, such as the list of all warning codes it can emit etc.

Reviewed By: dkgi

Differential Revision: D13427927

fbshipit-source-id: a427c34770940ff801773fe61cdbb2580d9da4ba