Add filter directory override to allow checking symlinked directories
Summary:
When passing in a source directory explicitly via the command line (no local config), pyre reasonably decides to consider any file that symlinks to outside of the given source directory to be a dependency and not report errors in it.

Usually, this boundary is determined by the location of the local configuration, or the location of the TARGETS file directory when a target is passed in with no configuration.

This feature is explicitly necessary for buck test integration, because the existing infrastructure creates a special typechecking target and passes the link tree for that typecheck target to mypy or pyre. We want to check the link tree, but can't rely on the default filtering logic (or everything will be filtered away, because they all symlink out to something living in fbcode/x not fbcode/buck-out/dev/gen/x). The original target doesn't even really exist; not filtering at all also doesn't work because we don't want to check all dependencies throughout fbcode. This was the best option I could think of.

Reviewed By: mrkmndz

Differential Revision: D13454180

fbshipit-source-id: cecfc411264c40bf6145060575a3090fb431c5dd