Differentiate syntax errors from system errors
Summary:
While parsing, syntax errors were summarily conflated with system
errors. This would cause problems in recognizing when an error is due
to the content of the file, as opposed to accessing the file.

This diff separates the two concerns, by changing the error message, from:
```
 ƛ Could not parse 3 files due to syntax errors!
```
to:
```
 ƛ Could not parse all files: 1 syntax error, 2 system errors!
```

The error message does not change if there is still one type of error
(either all errors are syntax errors, or all errors are system
errors).

Reviewed By: shannonzhu

Differential Revision: D13413046

fbshipit-source-id: 5331bf6bc63d010063d985330fe8df30154bc774