Fix inference of empty tuple type
Summary:
Noticed that printing the inferred type of empty tuple `()` literal
included type parameters: `typing.Tuple[typing.Any, ...]`. This is
incorrect because `Tuple[Any, ...]` and `Tuple[()]` are distinct annotations with different behaviors.

Ex.
`()` is a subtype of `Tuple[str, ...]` but `Tuple[Any, ...]` is not.
Nothing but an empty tuple is a subtype of `()` but most tuples are a subtype of `Tuple[Any, ...]`

MyPy docs: https://mypy.readthedocs.io/en/latest/builtin_types.html

Reviewed By: dark

Differential Revision: D13489220

fbshipit-source-id: 1e9d6e4dcfcfbd9daab1dab917837f18fd7eac9a