Breadcrumbs on taint flows
Summary:
Breadcrumbs are extra pieces of information we can attach to taint flows to make them easier to prioritize/filter.

This diff adds general support and also the `obscure` breadcrumb which indicates that the flow goes through an `unknown` function, i.e. one that we can't analyze and have no model for.

Reviewed By: dkgi

Differential Revision: D13442838

fbshipit-source-id: b1463120a6a65b514889a602a8a3ca8612e5bf02