Keep return constraints for callables when arity doesn't match up
Summary:
Resolving the return constraint for callables is still valuable if we can't match the arity - if there is a legitimate type error
here, Pyre will error with the arity issue. However, in cases such as default arguments and *args + **kwargs being present in the argument,
the existing logic is removing the type variable knowledge unnecessarily.

Reviewed By: mrkmndz

Differential Revision: D13510394

fbshipit-source-id: 46717aa391035e50b56f0031e0cd06d1cea75a0d