Fixed f-string preprocessing
The sequence `{{` or `}}` in an f-string escapes the corresponding curly brace. The interior expression should not be interpreted as an expression. So, I converted the f-string parsing from a regex matcher to a state machine, which should also be more amenable to further improvements e.g. expressions for format specifiers like `{value:{width}.{precision}}`.

Fixes #15.