Fixed f-string preprocessing (#128)
Summary:
The sequence `{{` or `}}` in an f-string escapes the corresponding curly brace. The interior expression should not be interpreted as an expression. So, I converted the f-string parsing from a regex matcher to a state machine, which should also be more amenable to further improvements e.g. expressions for format specifiers like `{value:{width}.{precision}}`.

Fixes #97 .
Pull Request resolved: https://github.com/facebook/pyre-check/pull/128

Reviewed By: shannonzhu

Differential Revision: D13529637

Pulled By: dark

fbshipit-source-id: f7c41679917bcc1a50393c9fe17377ccca0d2787