Fixing nested classes by introducing a type access fold
Summary: D13499641, D13532068, and  D13529896 all tried to get nested classes of aliases to work correctly, but turns out a more significant restructure was needed to get it to work.  Now we create a type by resolving aliases on increasing prefixes of a type, which allowed us to fix a lot of outstanding issues all at once.

Reviewed By: dkgi

Differential Revision: D13539985

fbshipit-source-id: de90e5ef4099e5a36e7772622d1069ae0853e0c7