Recursively collect elements of type to check against typeorder
Summary:
`Type.primitives` traverses the given annotation and returns whichever embedded primitives it finds. However, if we care about all of the "pieces" of that type (use-case, to check if any part of a given annotation is undefined), we need to also check the parametric bases.

Was fudging this before with a call to `Type.split` to extract the base as a primitive and check that, but want to make this recursive (nested parametric bases will be picked up) and centralized to use elsewhere (upcoming diff).

Naming suggestions welcome, really wasn't sure about the best terminology for this one.

Reviewed By: sinancepel

Differential Revision: D13551541

fbshipit-source-id: d9bd41ba82a6d579c4d1096fd8d3b912558740cc