pyre-upgrade emits multiple lines on long comments
Summary:
When `pyre-upgrade` has to emit a long comment, it might generate a
file that does not pass lint due to long lines. Manually re-flowing
that text in an IDE will break the comment itself, since Pyre will
consider only the first line as true comment; hence, the first line
would be incorrectly categorized as a superfluous error, and the
original error would still go unignored.

This diff changes the script so that it will reflow long comments
according to this format:

```
 # pyre: This is the format a long
 # pyre: comment would be reformatted
 # pyre-ignore[0]: to by the script.
```

Reviewed By: shannonzhu

Differential Revision: D13541849

fbshipit-source-id: aad242015f506daf8c9b319d83f8640ca280788c