Differentiate real typeshed builtins from helper stubs in testing
Summary:
Ran into an issue where `str` was not registered in the typeorder for a test, but when I added typeshed stubs to the environment being used, a lot of the pre-defined "extra" stubs we've stuffed into `builtins.pyi` started showing up and messing with the current test.

Splitting real typeshed stubs and our own convenience definitions, so that we can choose to use an environment with only the real stubs in certain tests if we choose to. If we need more builtin-like behavior or have helper stubs that are universally very useful we can keep adding to these, but ideally more cleaning can be done & we'll move towards not including them by default (and adding them manually when you want them) and defining any specific helpers you need locally.

Reviewed By: dark

Differential Revision: D13559156

fbshipit-source-id: 8330823b49569286ef7156432b4b47ba5ddfe7d6